Allow arbitrary number of permutations.
We have turned RandomLinearPermutations from an object into a class.
And MinHash is now a case class (instead of a type within a trait).

+++ src/main/scala/Constants.scala
@@ -7,5 +7,5 @@ object Constants {
   val hashValueByteSize: Int = 8
 
   val seed: Int = 1
-  val numPerm: Int = 256
+  val numPerm: Short = 256
 }
